+++ apps/website/src/components/demo/examples/card/default-card.tsx
@@ -5,7 +5,7 @@ export default function DefaultCard() {
         <Card.Root>
             <Card.Header>Basic Template</Card.Header>
             <Card.Body>This is a Basic Template</Card.Body>
-            <Card.Footer style={{ display: 'flex', justifyContent: 'center' }}>Footer</Card.Footer>
+            <Card.Footer>Footer</Card.Footer>
         </Card.Root>
     );
 }