+++ packages/codemod/src/transforms/internal/core/dropdown/__testfixtures__/dropdown-props-side-align.input.tsx
@@ -1,0 +1,15 @@
+// @ts-nocheck
+import { Dropdown } from '@goorm-dev/vapor-core';
+
+export default function App() {
+    return (
+        <Dropdown side="top" align="end">
+            <Dropdown.Trigger>Open</Dropdown.Trigger>
+            <Dropdown.Portal>
+                <Dropdown.Content>
+                    <Dropdown.Item>Item</Dropdown.Item>
+                </Dropdown.Content>
+            </Dropdown.Portal>
+        </Dropdown>
+    );
+}