+++ packages/core/src/components/floating-bar/index.ts
@@ -1,0 +1,1 @@
+export * as FloatingBar from './index.parts';