+++ packages/icons/src/components/basic-icons/IdeDocumentViewerOutlineIcon/index.ts
@@ -1,3 +1,0 @@
-import IdeDocumentViewerOutlineIcon from './IdeDocumentViewerOutlineIcon';
-
-export default IdeDocumentViewerOutlineIcon;